Rename FiatPluginEnterAmountScene file to be consistent
