+++ src/plugins/gui/scenes/FiatPluginEnterAmountScene.tsx
